moved interfaces to own files
