+++ LocalizationEditor/Localization/ClientApp/src/app/localization-edit-dialog/update-dialog-data.ts
@@ -1,0 +1,6 @@
+﻿import {LocalizationDataRow} from "../localization-table/localization-data-row";
+
+export interface UpdateDialogData {
+    locales: string[];
+    localizedString: LocalizationDataRow;
+}
