+++ LocalizationEditor/Localization/ClientApp/src/app/localization-table/localization-data-row.ts
@@ -1,0 +1,4 @@
+﻿export interface LocalizationDataRow {
+    group: string;
+    key: string;
+}
